Added current state of Unity Technologies BitBucket repo.
Added the files of the current AssetBundleDemo bitbucket from the Unity
Technologies BitBucket ( https://bitbucket.org/Unity-Technologies/assetbundledemo ).
The only file that was changed is that their readme was renamed to "README.OLD" so
as not to conflict with the new README.  Other than that files are unchanged as of
this commit.  (Mercurial tracking files were not imported to this repo.)